move project files to root level of repo
