--- MarsRoverChallenge_python/inputs/input1.txt
@@ -1,3 +1,0 @@
-94
-00 E
-M
