--- MarsRoverChallenge_python/inputs/input2.txt
@@ -1,3 +1,0 @@
-94
-00 N
-M
