--- MarsRoverChallenge_python/inputs/input4.txt
@@ -1,3 +1,0 @@
-94
-00 N
-R
