--- MarsRoverChallenge_python/inputs/input5.txt
@@ -1,3 +1,0 @@
-94
-00 W
-M
