--- MarsRoverChallenge_python/inputs/input6.txt
@@ -1,3 +1,0 @@
-94
-00 S
-M
