--- MarsRoverChallenge_python/inputs/input7.txt
@@ -1,3 +1,0 @@
-94
-94 E
-M
