--- MarsRoverChallenge_python/inputs/input8.txt
@@ -1,3 +1,0 @@
-94
-94 N
-M
