--- MarsRoverChallenge_python/inputs/input9.txt
@@ -1,3 +1,0 @@
-94
-94 N
-MRRMRR
