--- MarsRoverChallenge_python/main.py
@@ -1,6 +1,0 @@
-import sys
-import rover
-
-if __name__ == "__main__":
-	myRover = rover.Rover()
-	myRover.readInput(sys.argv[1])
