+++ inputs/input3.txt
@@ -1,0 +1,3 @@
+94
+00 N
+L
