+++ inputs/input7.txt
@@ -1,0 +1,3 @@
+94
+94 E
+M
