+++ inputs/input8.txt
@@ -1,0 +1,3 @@
+94
+94 N
+M
