+++ inputs/input9.txt
@@ -1,0 +1,3 @@
+94
+94 N
+MRRMRR
