+++ main.py
@@ -1,0 +1,6 @@
+import sys
+import rover
+
+if __name__ == "__main__":
+	myRover = rover.Rover()
+	myRover.readInput(sys.argv[1])
